Add files via upload
This is an updated version of Term project 2-repository profile. The original version is written in Feb 5th, the updated version is written in Feb 20th. By addressing the comment from professor Matthew S. Mayernik, the updated version elaborates on the potential reasoning behind GlobeData's SIP requirements, clarifying the use of the DDI standard, and highlighting the significance of the Dataverse software in supporting the repository's infrastructure.